+++ NEWS.md
@@ -1,3 +1,14 @@
+comtradr 0.0.1.9000
+===================
+
+### BUG FIXES
+
+* Add parameter `ssl_verify_peer` to the core package functions ([#1](https://github.com/ChrisMuir/comtradr/issues/1))
+
+### MINOR IMPROVEMENTS
+
+* Add unit tests via [testthat](https://github.com/hadley/testthat)
+
 comtradr 0.0.1 (2017-04-06)
 ===========================